+++ CHANGES.rst
@@ -8,6 +8,8 @@ Version 0.1.0
 * Added the :attr:`.User.location` attribute.
 * Removed support for :attr:`.User.country` and :attr:`.User.city` from
   :class:`.oauth1.Twitter` and :class:`.oauth2.GitHub` providers.
+* Removed support for :attr:`.User.link` and :attr:`.User.picture` from
+  :class:`.oauth1.Tumbler` provider.
 
 Version 0.0.13
 --------------